requirements.txt
# Claude Agent SDK
claude-agent-sdk>=0.1.0

# 核心依赖
pydantic>=2.0.0
pyyaml>=6.0.0

# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sse-starlette>=1.8.0

# 可选存储后端
# redis>=5.0.0      # pip install redis
asyncpg>=0.29.0     # PostgreSQL
